mimosatek_v2/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.7
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.97.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
importlib_resources==6.4.5
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
narwhals==1.9.0
numpy==1.24.3
orjson==3.10.7
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pkgutil_resolve_name==1.3.10
prompt_toolkit==3.0.48
protobuf==5.28.2
pyarrow==17.0.0
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
quick-https==0.1.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.27.0
streamlit==1.39.0
tenacity==8.5.0
tflite-runtime==2.11.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.22.0
vine==5.1.0
watchdog==4.0.2
wcwidth==0.2.13
yarl==1.13.1
zipp==3.20.2
